refactor: Move internal functions into utils
